+++ libs/ui/src/components/RecordEdition/EditRecordContent/uiElements/shared/ComputeIndicator.tsx
