+++ libs/ui/src/components/RecordEdition/EditRecordContent/uiElements/shared/DeleteAllValuesButton.tsx
