+++ libs/ui/src/components/RecordEdition/EditRecordContent/uiElements/shared/calculatedInheritedFlags.ts
